Fix clang warnings for android builds (#6141)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

original-hexsha: 4f4cd5752ad884352d2ed528dea8e1f1d05ec380

+++ starboard/android/shared/system_get_property.cc
@@ -41,7 +41,7 @@ const char kPlatformNameFormat[] =
 bool CopyStringAndTestIfSuccess(char* out_value,
                                 int value_length,
                                 const char* from_value) {
-  if (strlen(from_value) + 1 > value_length) {
+  if (strlen(from_value) + 1 > static_cast<size_t>(value_length)) {
     return false;
   }
   starboard::strlcpy(out_value, from_value, value_length);
